promote future into its own module
